backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx
requests
python-dotenv
aiofiles
mutagen
unidecode
thefuzz
python-levenshtein
ytmusicapi
passlib[bcrypt]
python-jose[cryptography]
python-multipart